For loop for partitioning + survival fraction
This commit adds a fct called 'test_surv_frac_diff_partitioning' for looping between partitioning factors (in this case we start with a small population in many droplets and multiply the initial population, but decrease the nr of droplets). The for loop is ready, but the survival fraction does not work 100%. I need to troubleshoot tomorrow.